requirements-vercel.txt
Django==5.2.1
whitenoise==6.9.0
dj-database-url==2.3.0
psycopg2-binary==2.9.10
python-dotenv==1.0.0
django-crispy-forms==2.0
django-allauth==0.58.0
reportlab==4.2.0
vercel-wsgi

# Development
django-extensions==3.2.3
